LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cache-storage-reporting-dedicated-worker.https-expected.txt


FAIL COEP support on DedicatedWorker. assert_equals: expected 1 but got 0
FAIL COEP-Report-Only support on DedicatedWorker. assert_equals: expected 1 but got 0
